requirements.txt
attr==0.3.2
beautifulsoup4==4.14.2
cloudscraper==1.2.71
ConfigParser==7.2.0
cryptography==46.0.3
faiss-cpu
fake_useragent==2.2.0
filelock==3.20.0
Flask==3.1.2
flask_cors==6.0.1
HTMLParser==0.0.2
ipython==9.7.0
ipywidgets==8.1.8
keyring==25.6.0
numpy==2.3.4
openai==2.7.1
protobuf==6.33.0
psycopg2_binary==2.9.10
pymongo==4.15.3
pyOpenSSL==25.3.0
ratelimit==2.2.1
redis==7.0.1
sentence_transformers==5.1.2
thread==2.0.5
tqdm==4.67.1
typing_extensions==4.15.0
